psr-4 fixes
closes #104
