+++ Website/htdocs/mpmanager/packages/inensus/bulk-registration/src/Exceptions/Handler.php
@@ -52,7 +52,7 @@ class Handler extends ExceptionHandler
     public function render($request, \Throwable $exception)
     {
         if ($exception instanceof CouldNotGeocode) {
-            new \GoogleMapsApiException(json_encode($exception->errors()));
+            new GoogleMapsApiException(json_encode($exception->errors()));
         }
 
         return parent::render($request, $exception);
