+++ Website/htdocs/mpmanager/packages/inensus/calin-smart-meter/src/Tests/Unit/TokenGenerationTest.php
