+++ Website/htdocs/mpmanager/packages/inensus/steama-meter/src/Exceptions/CronJobException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inensus\StemaMeter\Exceptions;
+namespace Inensus\SteamaMeter\Exceptions;
 
 class CronJobException extends \Exception
 {
