+++ Website/htdocs/mpmanager/packages/inensus/steama-meter/src/Exceptions/WrongCredentialsException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inensus\StemaMeter\Exceptions;
+namespace Inensus\SteamaMeter\Exceptions;
 
 class WrongCredentialsException extends \Exception
 {
