+++ Website/htdocs/mpmanager/packages/inensus/steama-meter/src/Tests/Feature/SendSms.php
@@ -1,23 +1,20 @@
 <?php
 
-namespace Inensus\SparkMeter\Test\Feature;
+namespace Inensus\SteamaMeter\Tests\Feature;
 
 use App\Models\Address\Address;
 use App\Models\MainSettings;
-use App\Models\Manufacturer;
-use App\Models\Meter\Meter;
-use App\Models\Meter\MeterTariff;
-use App\Models\Meter\MeterType;
 use App\Models\Person\Person;
 use App\Models\Sms;
 use App\Models\SmsBody;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Queue;
-use Inensus\SparkMeter\Models\SmCustomer;
-use Inensus\SparkMeter\Models\SmSite;
-use Inensus\SparkMeter\Models\SmSmsBody;
-use Inensus\SparkMeter\Models\SmSmsFeedbackWord;
+use Inensus\SteamaMeter\Models\SteamaCustomer;
+use Inensus\SteamaMeter\Models\SteamaSite;
+use Inensus\SteamaMeter\Models\SteamaSmsBody;
+use Inensus\SteamaMeter\Models\SteamaSmsFeedbackWord;
 use Tests\TestCase;
 use Tymon\JWTAuth\Facades\JWTAuth;
 
@@ -26,27 +23,10 @@ class SendSms extends TestCase
     use RefreshDatabase;
 
     /** @test */
-    public function isMeterResetFeedbackSend()
-    {
-        Queue::fake();
-        $this->withoutExceptionHandling();
-        $person = $this->initializeData()['customer'];
-        $user = factory(User::class)->create();
-
-        $data = [
-            'sender' => $person->addresses[0]->phone,
-            'message' => 'Reset',
-        ];
-        $response = $this->actingAs($user)->post('/api/sms', $data);
-        $response->assertStatus(201);
-        $smsCount = Sms::query()->first()->count();
-        $this->assertEquals(2, $smsCount);
-    }
-
-    /** @test */
     public function isMeterBalanceFeedbackSend()
     {
         Queue::fake();
+        Config::set('app.debug', false);
         $this->withoutExceptionHandling();
         $person = $this->initializeData()['customer'];
         $user = factory(User::class)->create();
@@ -55,7 +35,7 @@ class SendSms extends TestCase
             'message' => 'Balance',
         ];
         $response = $this->actingAs($user)->post('/api/sms', $data);
-        $response->assertStatus(200);
+        $response->assertStatus(201);
         $smsCount = Sms::query()->first()->count();
         $this->assertEquals(2, $smsCount);
     }
@@ -77,39 +57,9 @@ class SendSms extends TestCase
 
         // create person
         factory(Person::class)->create();
-        // create meter-tariff
-        factory(MeterTariff::class)->create();
-
-        // create meter-type
-        MeterType::query()->create([
-            'online' => 0,
-            'phase' => 1,
-            'max_current' => 10,
-        ]);
-
-        // create calin manufacturer
-        Manufacturer::query()->create([
-            'name' => 'Spark Meters',
-            'website' => 'https://www.sparkmeter.io/',
-            'api_name' => 'SparkMeterApi',
-        ]);
-
-        // create meter
-        Meter::query()->create([
-            'serial_number' => 'SM15R-01-000002F9',
-            'meter_type_id' => 1,
-            'in_use' => 1,
-            'manufacturer_id' => 1,
-        ]);
 
         // associate meter with a person
         $p = Person::query()->first();
-        $p->meters()->create([
-            'tariff_id' => 1,
-            'meter_id' => 1,
-            'connection_type_id' => 1,
-            'connection_group_id' => 1,
-        ]);
 
         // associate address with a person
         $address = Address::query()->make([
@@ -120,21 +70,19 @@ class SendSms extends TestCase
         $address->owner()->associate($p);
         $address->save();
 
-        SmSite::query()->create([
-            'site_id' => '1',
-            'mpm_mini_grid_id' => '1',
-            'thundercloud_url' => 'http://sparkapp-staging.spk.io:5010/api/v0',
-            'thundercloud_token' => '.eJwNw0EOgDAIBMC_cJbEhZbCW4yHbbX_f4JOMpfY4tjvGJonqM25NZNTmVb570STQx728h2zwZBnLloVyj2iFhwh9wfYAhKT.X2sZTg.JthNGNnFRaqEPqRJW8okuhzkucE',
-            'is_authenticated' => 1,
-            'is_online' => 1,
+        SteamaSite::query()->create([
+            'site_id' => 1,
+            'mpm_mini_grid_id' => 1,
         ]);
 
-        SmCustomer::query()->create([
-            'site_id' => '1',
-            'customer_id' => '6e70b116-4d0e-4975-b8f7-a948138cfa67',
+        SteamaCustomer::query()->create([
+            'site_id' => 1,
+            'customer_id' => 1,
+            'user_type_id' => 1,
             'mpm_customer_id' => $p->id,
-            'credit_balance' => 100,
-            'low_balance_limit' => 150,
+            'energy_price' => 100,
+            'account_balance' => 1000,
+            'low_balance_warning' => 150,
             'hash' => 'xxxxxxxxx',
         ]);
 
@@ -242,69 +190,45 @@ class SendSms extends TestCase
         }
         $smsBodies = [
             [
-                'reference' => 'SparkSmsLowBalanceHeader',
+                'reference' => 'SteamaSmsLowBalanceHeader',
                 'place_holder' => 'Dear [name] [surname],',
                 'variables' => 'name,surname',
                 'title' => 'Sms Header',
             ],
             [
-                'reference' => 'SparkSmsLowBalanceBody',
-                'place_holder' => 'your credit balance has reduced under [low_balance_limit],'
-                    .'your currently balance is [credit_balance]',
-                'variables' => 'low_balance_limit,credit_balance',
+                'reference' => 'SteamaSmsLowBalanceBody',
+                'place_holder' => 'your credit balance has reduced under [low_balance_warning],'
+                    .'your currently balance is [account_balance]',
+                'variables' => 'low_balance_warning,account_balance',
                 'title' => 'Low Balance Limit Notify',
             ],
             [
-                'reference' => 'SparkSmsBalanceFeedbackHeader',
+                'reference' => 'SteamaSmsBalanceFeedbackHeader',
                 'place_holder' => 'Dear [name] [surname],',
                 'variables' => 'name,surname',
                 'title' => 'Sms Header',
             ],
             [
-                'reference' => 'SparkSmsBalanceFeedbackBody',
-                'place_holder' => 'your currently balance is [credit_balance]',
-                'variables' => 'credit_balance',
+                'reference' => 'SteamaSmsBalanceFeedbackBody',
+                'place_holder' => 'your currently balance is [account_balance]',
+                'variables' => 'account_balance',
                 'title' => 'Balance Feedback',
             ],
             [
-                'reference' => 'SparkSmsMeterResetFeedbackHeader',
-                'place_holder' => 'Dear [name] [surname],',
-                'variables' => 'name,surname',
-                'title' => 'Sms Header',
-            ],
-            [
-                'reference' => 'SparkSmsMeterResetFeedbackBody',
-                'place_holder' => 'your meter, [meter_serial] has reset successfully.',
-                'variables' => 'meter_serial',
-                'title' => 'Meter Reset Feedback',
-            ],
-            [
-                'reference' => 'SparkSmsMeterResetFailedFeedbackBody',
-                'place_holder' => 'meter reset failed with [meter_serial].',
-                'variables' => 'meter_serial',
-                'title' => 'Meter Reset Failed Feedback',
-            ],
-            [
-                'reference' => 'SparkSmsMeterResetFeedbackFooter',
+                'reference' => 'SteamaSmsBalanceFeedbackFooter',
                 'place_holder' => 'Your Company etc.',
                 'variables' => '',
                 'title' => 'Sms Footer',
             ],
             [
-                'reference' => 'SparkSmsLowBalanceFooter',
-                'place_holder' => 'Your Company etc.',
-                'variables' => '',
-                'title' => 'Sms Footer',
-            ],
-            [
-                'reference' => 'SparkSmsBalanceFeedbackFooter',
+                'reference' => 'SteamaSmsLowBalanceFooter',
                 'place_holder' => 'Your Company etc.',
                 'variables' => '',
                 'title' => 'Sms Footer',
             ],
         ];
         collect($smsBodies)->each(function ($smsBody) {
-            SmSmsBody::query()->create([
+            SteamaSmsBody::query()->create([
                 'reference' => $smsBody['reference'],
                 'place_holder' => $smsBody['place_holder'],
                 'body' => $smsBody['place_holder'],
@@ -318,8 +242,7 @@ class SendSms extends TestCase
 
     private function addFeedBackKeys()
     {
-        SmSmsFeedbackWord::query()->create([
-            'meter_reset' => 'Reset',
+        SteamaSmsFeedbackWord::query()->create([
             'meter_balance' => 'Balance',
         ]);
     }
