+++ Website/htdocs/mpmanager/packages/inensus/swifta-payment-provider/src/Exceptions/SwiftaValidationFailedException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inensus\SwiftaPaymentProvider\Http\Exceptions;
+namespace Inensus\SwiftaPaymentProvider\Exceptions;
 
 class SwiftaValidationFailedException extends \Exception
 {
